bug fixes and reorganizing components
