+++ app/categories/[id]/CreateItem.jsx
@@ -3,7 +3,7 @@ import { useState } from "react";
 import { useUser } from "@/app/hooks/useUser";
 import { mutate } from "swr";
 import { createItem } from "@/app/lib/db";
-import ItemForm from "@/app/components/ItemForm";
+import ItemForm from "@/app/components/forms/ItemForm";
 import { notify } from "@/app/lib/handlers";
 
 const CreateItem = ({ data, close, mutateKey }) => {
