+++ app/components/CloudUploadButton.jsx
@@ -4,6 +4,7 @@ import { Upload } from "lucide-react";
 const CloudUploadButton = ({ handleUpload }) => {
   return (
     <CldUploadButton
+      multiple
       className="bg-black col-span-2 h-fit mt-8 py-3 rounded-xl font-semibold flex gap-1 justify-center items-center text-white"
       options={{
         multiple: true,
