+++ app/components/forms/AddItems.jsx
@@ -3,11 +3,10 @@ import FooterButtons from "../FooterButtons";
 import AddRemoveCard from "../AddRemoveCard";
 import Loading from "../Loading";
 import { addItems } from "@/app/lib/db";
-import { handleToggleSelect } from "@/app/lib/helpers";
+import { fetcher, handleToggleSelect } from "@/app/lib/helpers";
 import { notify } from "@/app/lib/handlers";
 import { mutate } from "swr";
 import useSWR from "swr";
-import { fetcher } from "@/app/lib/fetcher";
 import SearchFilter from "../SearchFilter";
 import { Suspense, useState } from "react";
 import { Loader, ScrollArea } from "@mantine/core";
