+++ app/components/index.jsx
@@ -1,8 +1,8 @@
 export { default as AddItems } from "./forms/AddItems";
 export { default as AddRemoveCard } from "./AddRemoveCard";
 export { default as AddRemoveSelector } from "./AddRemoveSelector";
-export { default as BackButton } from "./BackButton";
-export { default as BreadcrumbTrail } from "./BreadcrumbTrail";
+export { default as BackButton } from "./navigation/BackButton";
+export { default as BreadcrumbTrail } from "./navigation/BreadcrumbTrail";
 export { default as CategoryPill } from "./CategoryPill";
 export { default as CloudUploadButton } from "./CloudUploadButton";
 export { default as CloudUploadWidget } from "./CloudUploadWidget";
@@ -30,30 +30,29 @@ export { default as FavoriteFilterButton } from "./FavoriteFilterButton";
 export { default as FilterButton } from "./FilterButton";
 export { default as FilterPill } from "./FilterPill";
 export { default as FooterButtons } from "./FooterButtons";
-export { default as FormModal } from "./FormModal";
-export { default as GridLayout } from "./GridLayout";
-export { default as Header } from "./Header";
+export { default as GridLayout } from "./layout/GridLayout";
+export { default as Header } from "./navigation/Header";
 export { default as IconPill } from "./IconPill";
 export { default as ImageCarousel } from "./ImageCarousel";
 export { default as ImageLightbox } from "./ImageLightbox";
-export { default as ItemCardMasonry } from "./ItemCardMasonry";
+export { default as ItemCardMasonry } from "./navigation/ItemCardMasonry";
 export { default as ItemCountPill } from "./ItemCountPill";
 export { default as Loading } from "./Loading";
 export { default as LucideIcon } from "./LucideIcon";
-export { default as MasonryGrid } from "./MasonryGrid";
-export { default as MasonryContainer } from "./MasonryContainer";
-export { default as MobileMenu } from "./MobileMenu";
+export { default as MasonryGrid } from "./layout/MasonryGrid";
+export { default as MasonryContainer } from "./layout/MasonryContainer";
+export { default as MobileMenu } from "./navigation/MobileMenu";
 export { default as MultiSelect } from "./MultiSelect";
+export { default as NewCategory } from "./forms/NewCategory";
 export { default as NewContainer } from "./forms/NewContainer";
 export { default as NewLocation } from "./forms/NewLocation";
 export { default as PickerMenu } from "./forms/PickerMenu";
 export { default as SearchCard } from "./SearchCard";
 export { default as SearchFilter } from "./SearchFilter";
-export { default as Sidebar } from "./Sidebar";
-export { default as SortableItem } from "./SortableItem";
+export { default as Sidebar } from "./navigation/Sidebar";
 export { default as SquareItemCard } from "./SquareItemCard";
 export { default as Tooltip } from "./Tooltip";
-export { default as UniversalSearch } from "./UniversalSearch";
+export { default as UniversalSearch } from "./navigation/UniversalSearch";
 export { default as UpdateAvatar } from "./forms/UpdateAvatar";
 export { default as UpdateColor } from "./forms/UpdateColor";
 export { default as UpdateEmail } from "./forms/UpdateEmail";
