--- app/components/profile-server.jsx
@@ -1,20 +1,0 @@
-import { getSession } from "@auth0/nextjs-auth0";
-
-export default async function ProfileServer() {
-  const { user } = await getSession();
-
-  return (
-    user && (
-      <div>
-        <img src={user.picture} alt={user.name} width={100} height={100} />
-        <h2>{user.name}</h2>
-        <p>{user.email}</p>
-        {user ? (
-          <a href="/api/auth/logout">Log out {user?.name}?</a>
-        ) : (
-          <a href="/api/auth/logout">Log in</a>
-        )}
-      </div>
-    )
-  );
-}
