+++ app/containers/[id]/CreateItem.jsx
@@ -3,7 +3,7 @@ import { mutate } from "swr";
 import { notify } from "@/app/lib/handlers";
 import { useState } from "react";
 import { createItem } from "@/app/lib/db";
-import ItemForm from "@/app/components/ItemForm";
+import ItemForm from "@/app/components/forms/ItemForm";
 import { useUser } from "@/app/hooks/useUser";
 
 const CreateItem = ({ data, close, mutateKey }) => {
