+++ app/hooks/useUserColors.js
@@ -1,5 +1,5 @@
 import useSWR from "swr";
-import { fetcher } from "../lib/fetcher";
+import { fetcher } from "../lib/helpers";
 
 export const useUserColors = () => {
   const { data, error, isLoading } = useSWR(`/api/colors`, fetcher);
