--- app/lib/fetcher.js
@@ -1,1 +1,0 @@
-export const fetcher = (...args) => fetch(...args).then((res) => res.json());
