+++ app/locations/forms/NewItem.jsx
@@ -2,7 +2,7 @@
 import { notify } from "@/app/lib/handlers";
 import { useState } from "react";
 import { createItem } from "@/app/lib/db";
-import ItemForm from "@/app/components/ItemForm";
+import ItemForm from "@/app/components/forms/ItemForm";
 import { useUser } from "@/app/hooks/useUser";
 import { mutate } from "swr";
